+++ src/main/java/com/jihyoung/plant_disease_detection_web_spring/ai/dto/AiPredictResultResponse.java
@@ -1,12 +1,15 @@
 package com.jihyoung.plant_disease_detection_web_spring.ai.dto;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
 import com.jihyoung.plant_disease_detection_web_spring.pest.dto.info.PestInfoResponse;
+
+import java.math.BigDecimal;
 
 public record AiPredictResultResponse(
         PredictionStatus status,
         String cropName,
         String sickNameKor,
-        double confidence,
+        BigDecimal confidence,
         String message,
         PestInfoResponse pestInfo
 )
